pyproject.toml
[tool.poetry]
name = "loansforgood"
version = "0.1.0"
description = ""
authors = ["Lucas João de Oliveira"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
djangorestframework = "^3.14.0"
markdown = "^3.4.4"
django-filter = "^23.2"
psycopg2-binary = "^2.9.7"
requests = "^2.31.0"
redis = "^5.0.0"
celery = {extras = ["django"], version = "^5.3.3"}
logging = "^0.4.9.6"
prettyconf = "^2.2.1"
django-cors-headers = "^4.2.0"
django-prettyjson = "^0.4.1"
pytest = "^7.4.1"


[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"
pre-commit = "^3.3.3"
pytest-django = "^4.5.2"
factory-boy = "^3.3.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "loansforgood.settings"
